src/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==1.11.8
djangorestframework==3.6.2
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pygments==2.2.0
pytz==2018.4
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
